+++ frontend/src/App.js
@@ -7,6 +7,8 @@ import myInformation from './Pages/myInformation'
 import recommendation from './Pages/recommendation'
 import recommendation2 from './Pages/recommendation2'
 import recommendation3 from "./Pages/recommendation3";
+import recommendation_location from "./Pages/recommendation_location";
+
 import {Route, BrowserRouter as Router} from 'react-router-dom'
 function App () {
   return (
@@ -19,6 +21,7 @@ function App () {
             <Route path = "/nalot/recommendation" component = {recommendation}/>
             <Route path = "/nalot/recommendation2" component = {recommendation2}/>
             <Route path = "/nalot/recommendation3" component = {recommendation3}/>
+            <Route path = "/nalot/recommendation_location" component = {recommendation_location}/>
 
 
           </Router>
