+++ frontend/src/Pages/Main.js
@@ -149,6 +149,7 @@ function Main() {
 
         return year + month + day;
     }
+
     const hoursNow = () =>{
         let date = new Date();
 
@@ -163,6 +164,7 @@ function Main() {
         //console.log('recommend')
 
         axios.post('http://localhost:8080/weathers/forecasts',{
+
             "date":dateNow(), "time":"0500","nx":nx , "ny": ny
 
         }, {
@@ -178,7 +180,7 @@ function Main() {
                 console.log(response.data.temperatureMax)
                 console.log(response.data.temperatureMin)
 
-                history.push('/nalot/recommendation',{
+                history.push('/nalot/recommendation_location',{
                     "key": location.state.key,
                     "name": location.state.name,
                     "id": location.state.id,
@@ -207,66 +209,6 @@ function Main() {
 
     const logOut = () =>{
         console.log('logOut')
-    }
-
-    const submitLocation = () =>{
-        let s = document.getElementById("location")
-        let v = s.options[s.selectedIndex].value
-        console.log(v)
-        if (v === "seoul"){
-            nx = "60"
-            ny = "127"
-            loc = "서울특별시"
-        }else if (v === "busan"){
-            nx = "98"
-            ny = "76"
-            loc = "부산광역시"
-        }else if (v === "daegu"){
-            nx = "89"
-            ny = "90"
-            loc = "대구광역시"
-        }else{
-            nx = "60"
-            ny = "127"
-            loc = "서울특별시"
-        }
-
-        axios.post('http://localhost:8080/weathers/forecasts',{
-            "date":dateNow(), "time":"0500","nx":nx , "ny": ny
-
-        }, {
-            headers:{
-                'Content-Type':'application/json',
-                'Authorization':`${location.state.key}`
-
-            }
-        }).then(response=>{
-            history.push('/nalot/recommendation',{
-                "key": location.state.key,
-                "name": location.state.name,
-                "id": location.state.id,
-                "gender": location.state.gender,
-                "zone_code": location.state.zone_code,
-                "address_basic": location.state.address_basic,
-                "address_detail": location.state.address_detail,
-                "address_ground_number": location.state.address_ground_number,
-                "birth": location.state.birth,
-
-                "current":response.data.temperatureCurrent,
-                "max":response.data.temperatureMax,
-                "min":response.data.temperatureMin,
-
-                "nx" : nx,
-                "ny" : ny,
-                "loc" : loc
-
-            })
-        })
-    }
-
-    const valueChanged = () =>{
-        let s = document.getElementById("location")
-        let v  = s.options[s.selectedIndex].value
     }
 
     return (
@@ -320,10 +262,10 @@ function Main() {
                         <ListItemIcon><HomeIcon/></ListItemIcon>
                         <ListItemText primary={'홈'}/>
                     </ListItem>
-                        <ListItem button onClick={myInformation}>
-                            <ListItemIcon><AssignmentIndIcon /></ListItemIcon>
-                            <ListItemText primary={'내정보'} />
-                        </ListItem>
+                    <ListItem button onClick={myInformation}>
+                        <ListItemIcon><AssignmentIndIcon /></ListItemIcon>
+                        <ListItemText primary={'내정보'} />
+                    </ListItem>
                     <ListItem button onClick={myRecord}>
                         <ListItemIcon><InboxIcon /></ListItemIcon>
                         <ListItemText primary={'내 기록'} />
@@ -347,21 +289,6 @@ function Main() {
             </Drawer>
             <main className={classes.content}>
                 <div className={classes.toolbar} />
-                <div>
-                    지역 설정
-                </div>
-                <div>
-                    <select id="location" onChange={valueChanged}>
-                        <option value="seoul" defaultValue={true}>서울특별시</option>
-                        <option value="busan">부산광역시</option>
-                        <option value="daegu">대구광역시</option>
-                    </select>
-                </div>
-                <div>
-                    <button
-                        onClick={submitLocation}
-                    >옷 추천받기</button>
-                </div>
             </main>
         </div>
     );
