+++ frontend/src/Pages/index.js
@@ -7,6 +7,7 @@ import myInformation from"./myInformation"
 import recommendation from "./recommendation"
 import recommendation2 from "./recommendation2";
 import recommendation3 from "./recommendation3";
+import recommendation_location from "./recommendation_location";
 
 function Nalot({match}){
     return(
@@ -18,6 +19,8 @@ function Nalot({match}){
             <Route path = {`${match.path}/recommendation`} compoenent = {recommendation}/>
             <Route path = {`${match.path}/recommendation2`} component = {recommendation2}/>
             <Route path = {`${match.path}/recommendation3`} component = {recommendation3}/>
+            <Route path = {`${match.path}/recommendation_location`} component = {recommendation_location}/>
+
 
         </div>
     )
