+++ frontend/src/Pages/recommendation3.js
@@ -139,7 +139,7 @@ function recommendation3(){
 
     const recommendation = () =>{
         console.log('recommend')
-        history.push('/nalot/recommendation',{
+        history.push('/nalot/recommendation_location',{
             "key": location.state.key,
             "name": location.state.name,
             "id": location.state.id,
@@ -150,9 +150,6 @@ function recommendation3(){
             "address_ground_number": location.state.address_ground_number,
             "birth": location.state.birth,
 
-            "current":location.state.current,
-            "max":location.state.max,
-            "min":location.state.min,
         })
     }
 
@@ -165,7 +162,7 @@ function recommendation3(){
     }
 
     const submitClicked = () =>{
-        history.push('/nalot/recommendation',{
+        history.push('/nalot/recommendation_location',{
             "key": location.state.key,
             "name": location.state.name,
             "id": location.state.id,
@@ -176,9 +173,6 @@ function recommendation3(){
             "address_ground_number": location.state.address_ground_number,
             "birth": location.state.birth,
 
-            "current":location.state.current,
-            "max":location.state.max,
-            "min":location.state.min,
         })
     }
     const prevClicked = () =>{
@@ -194,8 +188,6 @@ function recommendation3(){
             "birth": location.state.birth,
 
             "current":location.state.current,
-            "max":location.state.max,
-            "min":location.state.min,
         })
     }
