+++ frontend/src/Pages/recommendation_location.js
@@ -20,8 +20,9 @@ import WbSunnyIcon from '@material-ui/icons/WbSunny';
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 import AccessibilityIcon from '@material-ui/icons/Accessibility';
 import AssignmentIndIcon from '@material-ui/icons/AssignmentInd';
+import HomeIcon from '@material-ui/icons/Home';
 import { useHistory, useLocation } from "react-router-dom";
-import HomeIcon from "@material-ui/icons/Home";
+import axios from'axios'
 
 const drawerWidth = 240;
 
@@ -87,13 +88,17 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-
-function recommendation2(){
+function recommendation_location(){
     const classes = useStyles();
     const theme = useTheme();
     const history = useHistory();
     const [open, setOpen] = React.useState(false);
     const location = useLocation();
+    let nx, ny, loc;
+
+    nx = "60"
+    ny = "127"
+    loc = "서울특별시"
 
     const handleDrawerOpen = () => {
         setOpen(true);
@@ -118,6 +123,7 @@ function recommendation2(){
     }
 
     const myInformation = () =>{
+
         history.push('/nalot/myinformation',{
             "key": location.state.key,
             "name": location.state.name,
@@ -135,20 +141,61 @@ function recommendation2(){
     const myRecord = () =>{
         console.log('myRecord')
     }
+    const dateNow = () =>{
+        let date = new Date();
+        let year = date.getFullYear();
+        let month = ("0" + (1 + date.getMonth())).slice(-2);
+        let day = ("0" + date.getDate()).slice(-2);
+
+        return year + month + day;
+    }
+
+    const hoursNow = () =>{
+        let date = new Date();
+
+        let hours = date.getHours();
+        console.log((hours-1+'00'))
+        //조건문 사용해서 02일 00시 -> 01시 23시로 바꿀것
+        //매시 30분 마다 업뎃 -> getMinutes사용해서 30분ㅇ ㅣ후면 아워리턴 이전이면 아ㅜ어 -1 리턴
+        return (hours-1) +'00';
+    }
 
     const recommendation = () =>{
-        console.log('recommend')
-        history.push('/nalot/recommendation_location',{
-            "key": location.state.key,
-            "name": location.state.name,
-            "id": location.state.id,
-            "gender": location.state.gender,
-            "zone_code": location.state.zone_code,
-            "address_basic": location.state.address_basic,
-            "address_detail": location.state.address_detail,
-            "address_ground_number": location.state.address_ground_number,
-            "birth": location.state.birth,
+
+        axios.post('http://localhost:8080/weathers/forecasts',{
+
+            "date":dateNow(), "time":"0500","nx":nx , "ny": ny
+
+        }, {
+            headers:{
+                'Content-Type':'application/json',
+                'Authorization':`${location.state.key}`
+
+            }
         })
+            .then(response=>{
+                console.log(response.data)
+                console.log(response.data.temperatureCurrent)
+
+                history.push('/nalot/recommendation_location',{
+                    "key": location.state.key,
+                    "name": location.state.name,
+                    "id": location.state.id,
+                    "gender": location.state.gender,
+                    "zone_code": location.state.zone_code,
+                    "address_basic": location.state.address_basic,
+                    "address_detail": location.state.address_detail,
+                    "address_ground_number": location.state.address_ground_number,
+                    "birth": location.state.birth,
+
+                    "current":response.data.temperatureCurrent,
+
+                    "nx" : nx,
+                    "ny" : ny,
+                    "loc" : loc
+
+                })
+            })
     }
 
     const weatherNow = () =>{
@@ -159,40 +206,64 @@ function recommendation2(){
         console.log('logOut')
     }
 
-    const prevClicked = () =>{
-        history.push('/nalot/recommendation',{
-            "key": location.state.key,
-            "name": location.state.name,
-            "id": location.state.id,
-            "gender": location.state.gender,
-            "zone_code": location.state.zone_code,
-            "address_basic": location.state.address_basic,
-            "address_detail": location.state.address_detail,
-            "address_ground_number": location.state.address_ground_number,
-            "birth": location.state.birth,
+    const submitLocation = () =>{
+        let s = document.getElementById("location")
+        let v = s.options[s.selectedIndex].value
+        console.log(v)
+        if (v === "seoul"){
+            nx = "60"
+            ny = "127"
+            loc = "서울특별시"
+        }else if (v === "busan"){
+            nx = "98"
+            ny = "76"
+            loc = "부산광역시"
+        }else if (v === "daegu"){
+            nx = "89"
+            ny = "90"
+            loc = "대구광역시"
+        }else{
+            nx = "60"
+            ny = "127"
+            loc = "서울특별시"
+        }
 
-            "current":location.state.current,
+        axios.post('http://localhost:8080/weathers/forecasts',{
+            "date":dateNow(), "time":"0500","nx":nx , "ny": ny
+
+        }, {
+            headers:{
+                'Content-Type':'application/json',
+                'Authorization':`${location.state.key}`
+
+            }
+        }).then(response=>{
+            history.push('/nalot/recommendation',{
+                "key": location.state.key,
+                "name": location.state.name,
+                "id": location.state.id,
+                "gender": location.state.gender,
+                "zone_code": location.state.zone_code,
+                "address_basic": location.state.address_basic,
+                "address_detail": location.state.address_detail,
+                "address_ground_number": location.state.address_ground_number,
+                "birth": location.state.birth,
+
+                "current":response.data.temperatureCurrent,
+
+                "nx" : nx,
+                "ny" : ny,
+                "loc" : loc
+
+            })
         })
     }
 
-    const nextClicked = () =>{
-        history.push('/nalot/recommendation3',{
-            "key": location.state.key,
-            "name": location.state.name,
-            "id": location.state.id,
-            "gender": location.state.gender,
-            "zone_code": location.state.zone_code,
-            "address_basic": location.state.address_basic,
-            "address_detail": location.state.address_detail,
-            "address_ground_number": location.state.address_ground_number,
-            "birth": location.state.birth,
-
-            "current":location.state.current,
-        })
+    const valueChanged = () =>{
+        let s = document.getElementById("location")
+        let v  = s.options[s.selectedIndex].value
     }
-
     return(
-
         <div className={classes.root}>
             <CssBaseline />
             <AppBar
@@ -271,49 +342,23 @@ function recommendation2(){
             <main className={classes.content}>
                 <div className={classes.toolbar} />
                 <div>
-                    <span>선택한 의상</span>
-                    &nbsp;
-                    &nbsp;
-                    &nbsp;
-                    <span>톤온톤 매치</span>
-                </div>
-                <br></br>
-                <div>
-                    <button>선택한 의상</button>
-                    &nbsp;
-                    &nbsp;
-                    &nbsp;
-                    <button>톤온톤 매치</button>
-                </div>
-                <br></br>
-                <div>
-                    <span>선택한 색상</span>
-                    &nbsp;
-                    &nbsp;
-                    &nbsp;
-                    <span>톤인톤 매치</span>
+                    지역 설정
                 </div>
                 <div>
-                    <button>선택한 색상</button>
-                    &nbsp;
-                    &nbsp;
-                    &nbsp;
-                    <button>톤인톤 매치</button>
+                    <select id="location" onChange={valueChanged}>
+                        <option value="seoul" defaultValue={true}>서울특별시</option>
+                        <option value="busan">부산광역시</option>
+                        <option value="daegu">대구광역시</option>
+                    </select>
                 </div>
                 <div>
-                <button
-                    onClick={prevClicked}
-                >이전</button>
-                <button
-                onClick={nextClicked}
-                >다음</button>
+                    <button
+                        onClick={submitLocation}
+                    >옷 추천받기</button>
                 </div>
             </main>
-
         </div>
-
-
-    );
+    )
 }
 
-export default recommendation2;
+export default recommendation_location;
